:recycle: refactor: move programmatic API from bin/compute-version to lib/compute-version
