+++ lib/commits.js
@@ -30,7 +30,7 @@ export function normalizeSubject(subject = "") {
     .trim();
 }
 
-export function cleanSubject(subject) {
+export function cleanSubject(subject = "") {
   const s = normalizeSubject(subject).replace(COMMIT_RE, "");
   return s ? s.charAt(0).toUpperCase() + s.slice(1) : s;
 }
@@ -99,12 +99,12 @@ export function categorizeCommits(commits = []) {
   for (const commit of commits) {
     const bump = detectBumpType(commit);
     const rawSubject = normalizeSubject(commit.subject);
-    const cleanSubject = cleanSubject(commit.subject);
+    const clearedSubject = cleanSubject(commit.subject);
 
     const view = {
       ...commit,
       rawSubject,
-      subject: cleanSubject,
+      subject: clearedSubject,
     };
 
     if (bump === "major") {
