+++ lib/versioning.js
@@ -22,12 +22,12 @@ import { COMMIT_RE } from "./constants.js";
  * @returns {'major'|'minor'|'patch'|'none'} The semantic version bump type.
  */
 export function detectBumpType({ subject, body }) {
-  const cleanSubject = normalizeSubject(subject);
+  const normalizedSubject = normalizeSubject(subject);
 
   // Ignore revert commits entirely
-  if (/^revert\b/i.test(cleanSubject)) return "none";
+  if (/^revert\b/i.test(normalizedSubject)) return "none";
 
-  const match = cleanSubject.match(COMMIT_RE);
+  const match = normalizedSubject.match(COMMIT_RE);
 
   const breaking =
     /BREAKING CHANGE/i.test(body) || (match && match[3] === "!");
